Add files via upload
Problem Statement: - To Build the Machine Learning Model using algorithm to predict the sale price for each house.
Key role: In this project I Worked on Imbalance Dataset, Removing Outliers, and creating a machine-learning model with linear regression on the housing dataset. This model will aid us in making better real estate decisions by making house price predictions in that area For this, we will carry out data exploration to better understand the data and will require preprocessing to improve the model's accuracy.